web_api/requirements.txt
pyyaml~=5.3.1
tinydb~=4.2.0
flask~=1.1.2
flask-restplus
werkzeug==0.16.1
flask-cors